+++ examples/ERP/plot_embedding_MEG.py
@@ -1,9 +1,10 @@
 """
 =====================================================================
-Embedding ERP MEG data in 2D Euclidean space with Laplacian Eigenmaps
-=====================================================================
+Embedding ERP MEG data in 2D Euclidean space
+============================================
 
-Spectral embedding via Laplacian Eigenmaps of a set of ERP data.
+Embeddings via Laplacian Eigenmaps and Riemannian locally linear
+embedding of a set of ERP data.
 
 """
 # Authors: Pedro Rodrigues <pedro.rodrigues01@gmail.com>
@@ -11,7 +12,7 @@ Spectral embedding via Laplacian Eigenmaps of a set of ERP data.
 # License: BSD (3-clause)
 
 from pyriemann.estimation import XdawnCovariances
-from pyriemann.embedding import Embedding
+from pyriemann.embedding import Embedding, RiemannLLE
 
 import mne
 from mne import io
@@ -50,7 +51,7 @@ X = epochs.get_data()
 y = epochs.events[:, -1]
 
 ###############################################################################
-# Embedding the Xdawn covariance matrices with Laplacian Eigenmaps
+# Embedding of Xdawn covariance matrices
 
 nfilter = 4
 xdwn = XdawnCovariances(estimator='scm', nfilter=nfilter)
@@ -58,11 +59,15 @@ split = train_test_split(X, y, train_size=0.25, random_state=42)
 Xtrain, Xtest, ytrain, ytest = split
 covs = xdwn.fit(Xtrain, ytrain).transform(Xtest)
 
+###############################################################################
+# Spectral Embedding (SE)
+# -----------------------
+
 lapl = Embedding(metric='riemann', n_components=2)
 embd = lapl.fit_transform(covs)
 
 ###############################################################################
-# Plot the three first components of the embedded points
+# Plot the two first components of the SE embedded points
 
 fig, ax = plt.subplots(figsize=(7, 8), facecolor='white')
 
@@ -75,6 +80,30 @@ ax.set_ylabel(r'$\varphi_2$', fontsize=16)
 ax.set_title('Spectral embedding of ERP recordings', fontsize=16)
 ax.set_xticks([-1.0, -0.5, 0.0, +0.5, 1.0])
 ax.set_yticks([-1.0, -0.5, 0.0, +0.5, 1.0])
+ax.grid(False)
+ax.legend()
+plt.show()
+
+###############################################################################
+# Riemannian Locally Linear Embedding (RLLE)
+# ------------------------------------------
+
+rlle = RiemannLLE(n_components=2)
+embd = rlle.fit_transform(covs)
+
+###############################################################################
+# Plot the two first components of the RLLE embedded points
+
+fig, ax = plt.subplots(figsize=(7, 8), facecolor='white')
+
+for cond, label in event_id.items():
+    idx = (ytest == label)
+    ax.scatter(embd[idx, 0], embd[idx, 1], s=36, label=cond)
+
+ax.set_xlabel(r'$\varphi_1$', fontsize=16)
+ax.set_ylabel(r'$\varphi_2$', fontsize=16)
+ax.set_title('Riemannian Locally Linear Embedding of ERP recordings',
+             fontsize=16)
 ax.grid(False)
 ax.legend()
 plt.show()